refactor(localization): share AI translation scope helpers
